requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.3
Django==3.1.4
django-allauth==0.44.0
django-bulk-update==2.2.0
django-cors-headers==3.6.0
django-environ==0.4.5
django-heroku==0.3.1
django-reset-migrations==0.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-jwt-2fa==0.3.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.3
websocket-client==1.2.3
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.4.0